pyproject.toml
[project]
name = "ontology-reasoning-system"
version = "0.1.0"
description = "Ontology 아키텍처 기반 지식 그래프 추론 시스템 - ToG 3.0 MACER + LangGraph + Neo4j"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["ontology", "knowledge-graph", "neo4j", "langgraph", "reasoning"]

dependencies = [
    # LangChain Ecosystem
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-neo4j>=0.2.0",
    "langchain-text-splitters>=0.3.0",

    # LangGraph for Orchestration
    "langgraph>=0.2.0",
    "langgraph-cli[inmem]>=0.1.0",
    "langgraph-sdk>=0.1.0",

    # Neo4j
    "neo4j>=5.20.0",

    # FastAPI & Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.9",

    # Document Processing
    "PyMuPDF>=1.24.0",
    "pypdf>=4.0.0",
    "PyYAML>=6.0.0",
    "ijson>=3.2.0",  # Streaming JSON parser for large files
    "beautifulsoup4>=4.12.0",  # HTML parsing
    "lxml>=5.0.0",  # XML parsing
    "defusedxml>=0.7.0",  # Secure XML parsing

    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "aiofiles>=24.1.0",

    # Observability
    "structlog>=24.4.0",

    # LangGraph Checkpointing (PostgreSQL)
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary]>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
